refactor(ui): extract HealthDisplayMixin for wound table rendering
Move the four wound-table render variants plus update_wound_penalties
and get_health_rank into l5r/ui/health.py. The mixin still reads
self.wounds (built by Tab 1) and the L5RCMSettings.app.health_method
preference exactly as before.

main.py: 2361 -> 2300 lines.
