+++ l5r/main.py
@@ -47,6 +47,7 @@ import l5r.dialogs as dialogs
 
 from l5r.l5rcmcore import *
 from l5r.ui.advise import AdviseMixin
+from l5r.ui.health import HealthDisplayMixin
 from l5r.ui.helpers import (
     new_horiz_line,
     new_item_groupbox,
@@ -59,7 +60,7 @@ from l5r.util import log
 from l5r.util.settings import L5RCMSettings
 
 
-class L5RMain(AdviseMixin, NicebarMixin, L5RCMCore):
+class L5RMain(AdviseMixin, HealthDisplayMixin, NicebarMixin, L5RCMCore):
 
     default_size = QtCore.QSize(820, 720)
     default_point_size = 8.25
@@ -2029,65 +2030,6 @@ class L5RMain(AdviseMixin, NicebarMixin, L5RCMCore):
             v.resizeColumnsToContents()
             v.setVisible(True)
 
-    def update_wound_penalties(self):
-        WOUND_PENALTIES_NAMES = [
-            self.tr("Healthy"),
-            self.tr("Nicked"),
-            self.tr("Grazed"),
-            self.tr("Hurt"),
-            self.tr("Injured"),
-            self.tr("Crippled"),
-            self.tr("Down"),
-        ]
-
-        for i in reversed(range(0, 7)):
-            if i < 7:
-                penalty = api.rules.get_wound_penalties(i)
-                text = u"{0} (+{1})".format(WOUND_PENALTIES_NAMES[i], penalty)
-            else:
-                text = WOUND_PENALTIES_NAMES[i]
-            self.wounds[i][0].setText(text)
-
-        # TODO Create a generate mechanism for data-pack defined bonus to penalties.
-        # TODO toku bushi school removes some penalties
-
-    def display_health(self):
-        settings = L5RCMSettings()
-        method = settings.app.health_method
-        if method == 'default':
-            self.display_health_default()
-        elif method == 'wounds':
-            self.display_total_wounds()
-        else:
-            self.display_health_stacked()
-
-
-    def display_health_default(self):
-        wounds_table = api.rules.get_wounds_table()
-        if not wounds_table:
-            return        
-        for i, (i_inc, i_total, i_stacked, i_inc_wounds, i_total_wounds, i_stacked_wounds) in enumerate(wounds_table):
-            self.wounds[i][1].setText(str(i_inc))
-            self.wounds[i][2].setText(str(i_inc_wounds) if i_inc_wounds else '')
-
-
-    def display_health_stacked(self):
-        wounds_table = api.rules.get_wounds_table()
-        if not wounds_table:
-            return
-        for i, (i_inc, i_total, i_stacked, i_inc_wounds, i_total_wounds, i_stacked_wounds) in enumerate(wounds_table):
-            self.wounds[i][1].setText(str(i_total))
-            self.wounds[i][2].setText(str(i_total_wounds) if i_total_wounds else '')
-
-
-    def display_total_wounds(self):
-        wounds_table = api.rules.get_wounds_table()
-        if not wounds_table:
-            return
-        for i, (i_inc, i_total, i_stacked, i_inc_wounds, i_total_wounds, i_stacked_wounds) in enumerate(wounds_table):
-            self.wounds[i][1].setText(str(i_stacked))
-            self.wounds[i][2].setText(str(i_stacked_wounds) if i_stacked_wounds else '')
-
     def closeEvent(self, ev):
         # update interface last time, to set unsaved states
         self.update_from_model()
@@ -2231,9 +2173,6 @@ class L5RMain(AdviseMixin, NicebarMixin, L5RCMCore):
     def create_new_character(self):
         self.sink1.new_character()
         self.pc.unsaved = False
-
-    def get_health_rank(self, idx):
-        return self.wounds[idx][1].text()
 
 # MAIN ###
